Status form + RelationType Form + Remove Management form for Ticket, Task, Member + TimeTracker

+++ src/Harproject/AppBundle/Controller/Management/GroupController.php
@@ -187,6 +187,10 @@ class GroupController extends Controller
             throw $this->createNotFoundException('Unable to find Group entity.');
         }
 
+        if ($entity->getIsLocked()) {
+            throw $this->createNotFoundException('Unable to delete Group entity.');
+        }
+        
         $em->remove($entity);
         $em->flush();
